extract ProposalForm and ProposalListItem components
